% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basinFishTotals.R
\name{basinFishTotals}
\alias{basinFishTotals}
\title{Summarize watershed fish totals}
\usage{
basinFishTotals(db, watershed.code)
}
\arguments{
\item{db}{The exported database}

\item{watershed.code}{The watershed.id of the basin to summarize}
}
\value{
A table of total catch by species per watershed
}
\description{
Provide a report on watershed by fish totals to date
}
\details{
Summarize species totals caught per watershed. Any values with an NA in the
weight_kg field will be automatically dropped, as will any values without an
associated scientific name. Could possibly lump those in with Fish sp.
}
\examples{
#load the data
data(BDER_20220110)

#note that tip states comes first here!
result <- basinFishTotals(db=BDER_20220110, watershed.code="W070408")

}
\author{
Eliot Miller & Cullen Hanks
}
