% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BDER_20220110}
\alias{BDER_20220110}
\title{Data export provided by Cullen.}
\format{
csv
}
\source{
\url{https://ictio.org/}
}
\usage{
BDER_20220110
}
\description{
Provided 7Feb2022.
}
\keyword{datasets}
