% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFishTotals.R
\name{userFishTotals}
\alias{userFishTotals}
\title{Summarize user fish totals}
\usage{
userFishTotals(db, user.id)
}
\arguments{
\item{db}{The exported database}

\item{user.id}{The user.id of the user to summarize}
}
\value{
A table of total catch by species per user
}
\description{
Provide a report on user by fish totals to date
}
\details{
Summarize species totals caught per user. Any values with an NA in the
weight_kg field will be automatically dropped, as will any values without an
associated scientific name. Could possibly lump those in with Fish sp.
}
\examples{
#load the data
data(BDER_20220110)

#note that tip states comes first here!
result <- userFishTotals(db=BDER_20220110, user.id="USER2385074")

}
\author{
Eliot Miller & Cullen Hanks
}
